Update sayBatteryLevel function to reflect the output type of readBatteryLevel function.

+++ components/custom_synth.h
@@ -79,7 +79,7 @@ void sayBatteryLevel();
 // Brief: Vocalize given number parameter.
 // Param:  number Number to vocalize.
 // Return: none 
-void sayBatteryPercent(float number);
+void sayBatteryPercent(int number);
 
 
 void sayWelocome();
